refactor(assistant): move tool/trace/skill telemetry stores out of memory/ into telemetry/
