+++ assistant/src/__tests__/audit-log-rotation.test.ts
@@ -12,12 +12,12 @@ mock.module("../util/logger.js", () => ({
     }),
 }));
 
+import { getSqlite } from "../persistence/db-connection.js";
+import { initializeDb } from "../persistence/db-init.js";
 import {
   getRecentInvocations,
   rotateToolInvocations,
-} from "../memory/tool-usage-store.js";
-import { getSqlite } from "../persistence/db-connection.js";
-import { initializeDb } from "../persistence/db-init.js";
+} from "../telemetry/tool-usage-store.js";
 import { resetDbForTesting } from "./db-test-helpers.js";
 
 // ---------------------------------------------------------------------------
