+++ assistant/src/__tests__/conversation-app-control-instantiation.test.ts
@@ -202,7 +202,7 @@ mock.module("../skills/catalog-cache.js", () => ({
 mock.module("../skills/install-meta.js", () => ({
   readInstallMeta: () => null,
 }));
-mock.module("../memory/skill-loaded-events-store.js", () => ({
+mock.module("../telemetry/skill-loaded-events-store.js", () => ({
   recordSkillLoadedEvent: () => {},
 }));
