+++ assistant/src/__tests__/conversation-process-app-control-preactivation.test.ts
@@ -62,7 +62,7 @@ mock.module("../memory/canonical-guardian-store.js", () => ({
   listPendingRequestsByConversationScope: () => [],
 }));
 
-mock.module("../memory/trace-event-store.js", () => ({
+mock.module("../telemetry/trace-event-store.js", () => ({
   persistTraceEvent: () => {},
   getMaxSequence: () => 0,
 }));
