+++ assistant/src/__tests__/trace-emitter.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, mock } from "bun:test";
 
 // Mock the persistence layer so tests are isolated from the database.
 // getMaxSequence returning -1 means "no persisted events" → sequence starts at 0.
-mock.module("../memory/trace-event-store.js", () => ({
+mock.module("../telemetry/trace-event-store.js", () => ({
   getMaxSequence: () => -1,
   persistTraceEvent: () => {},
 }));
