+++ assistant/src/telemetry/skill-loaded-events-store.test.ts
