+++ assistant/src/telemetry/skill-loaded-events-store.ts
