+++ assistant/src/telemetry/tool-executed-events-store.test.ts
