+++ assistant/src/telemetry/tool-executed-events-store.ts
