+++ assistant/src/telemetry/tool-usage-store.ts
