+++ assistant/src/telemetry/trace-event-store.ts
